Avoid possible exception while processing exception case
- Reuse ModuleSignalSentry for this case.
- Keep past behavior where exception in post signal during begin also resulted in module end not being called.
